+++ kinto-remote-settings/src/kinto_remote_settings/signer/generate_keypair.py
@@ -15,6 +15,9 @@ def generate_keypair(private_key_location, public_key_location):
 
 if __name__ == "__main__":  # pragma: no cover
     if len(sys.argv) != 3:
-        print("Usage: python -m signer.generate_keypair " "{public_key} {private_key}")
+        print(
+            "Usage: python -m kinto_remote_settings.signer.generate_keypair "
+            "{private_key} {public_key}"
+        )
         sys.exit(0)
     generate_keypair(sys.argv[1], sys.argv[2])