+++ CHANGELOG.rst
@@ -26,6 +26,7 @@ the version control of each dependency.
 - Default value for ``kinto.signer.resources`` is now ``/buckets/main-workspace -> /buckets/main-preview -> /buckets/main``
 - Now ship default configuration files to run a local instance of Remote Settings out of the box (see the *Setup a Local Server* tutorial)
 - The container can now run by just setting the ``KINTO_INI`` environment variable ``docker run -e KINTO_INI=config/testing.ini mozilla/remote-settings``
+- The container now runs without custom configuration with ``docker run mozilla/remote-settings``
 
 **Bug Fixes**